ansible_collections/ansiblesharp/az/requirements.txt
azure-cli-core
azure-common
azure-core
azure-identity
azure-mgmt-resource